+++ include/engine/raster/OpenGLRasterizer.h
@@ -124,6 +124,8 @@ namespace engine::raster {
 
     Rasterizer::AttachmentLoadOp colorLoadOp() const;
     void setColorLoadOp(Rasterizer::AttachmentLoadOp op);
+    const Buffer<Colord>* colorLoadSource() const;
+    void setColorLoadSource(const Buffer<Colord>* buffer);
     Rasterizer::AttachmentStoreOp colorStoreOp() const;
     void setColorStoreOp(Rasterizer::AttachmentStoreOp op);
 
@@ -135,6 +137,8 @@ namespace engine::raster {
     void setDepthClearValue(double value);
     Rasterizer::AttachmentLoadOp depthLoadOp() const;
     void setDepthLoadOp(Rasterizer::AttachmentLoadOp op);
+    const Buffer<double>* depthLoadSource() const;
+    void setDepthLoadSource(const Buffer<double>* buffer);
     Rasterizer::AttachmentStoreOp depthStoreOp() const;
     void setDepthStoreOp(Rasterizer::AttachmentStoreOp op);
     bool depthWriteEnabled() const;
@@ -171,6 +175,8 @@ namespace engine::raster {
     void setStencilClearValue(std::uint8_t value);
     Rasterizer::AttachmentLoadOp stencilLoadOp() const;
     void setStencilLoadOp(Rasterizer::AttachmentLoadOp op);
+    const Buffer<std::uint8_t>* stencilLoadSource() const;
+    void setStencilLoadSource(const Buffer<std::uint8_t>* buffer);
     Rasterizer::AttachmentStoreOp stencilStoreOp() const;
     void setStencilStoreOp(Rasterizer::AttachmentStoreOp op);
     std::uint8_t stencilWriteMask() const;
@@ -225,11 +231,13 @@ namespace engine::raster {
     bool m_scissorTestEnabled{false};
     Recti m_scissorRect;
     Rasterizer::AttachmentLoadOp m_colorLoadOp{Rasterizer::AttachmentLoadOp::Clear};
+    const Buffer<Colord>* m_loadColorAttachment{nullptr};
     Rasterizer::AttachmentStoreOp m_colorStoreOp{Rasterizer::AttachmentStoreOp::Store};
     Rasterizer::DepthFunc m_depthFunc{Rasterizer::DepthFunc::Less};
     double m_depthBias{0.0};
     double m_depthClearValue{std::numeric_limits<double>::infinity()};
     Rasterizer::AttachmentLoadOp m_depthLoadOp{Rasterizer::AttachmentLoadOp::Clear};
+    const Buffer<double>* m_loadDepthAttachment{nullptr};
     Rasterizer::AttachmentStoreOp m_depthStoreOp{Rasterizer::AttachmentStoreOp::Store};
     bool m_depthWriteEnabled{true};
     std::uint8_t m_colorWriteMask{Rasterizer::ColorWriteAll};
@@ -248,6 +256,7 @@ namespace engine::raster {
     std::uint8_t m_stencilMask{0xff};
     std::uint8_t m_stencilClearValue{0};
     Rasterizer::AttachmentLoadOp m_stencilLoadOp{Rasterizer::AttachmentLoadOp::Clear};
+    const Buffer<std::uint8_t>* m_loadStencilAttachment{nullptr};
     Rasterizer::AttachmentStoreOp m_stencilStoreOp{Rasterizer::AttachmentStoreOp::Store};
     std::uint8_t m_stencilWriteMask{0xff};
     Rasterizer::StencilOp m_stencilFailOp{Rasterizer::StencilOp::Keep};